+++ nature/opts/init.lua
@@ -25,7 +25,8 @@ local defaultOpts = {
 	greeting = string.format([[Welcome to {magenta}Hilbish{reset}, {cyan}%s{reset}.
 The nice lil shell for {blue}Lua{reset} fanatics!
 ]], hilbish.user),
-	motd = true
+	motd = true,
+	insensitive = false
 }
 
 for optsName, default in pairs(defaultOpts) do